requirements.txt
psycopg2-binary==2.9.6
requests==2.31.0